+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 #### v1.1.0
 
+- File name changed to catherinefbhd.suprx
 - OSD indicating patch success or failure